+++ src/components/toasts-list/components/transaction-toast/transaction-toast.tsx
@@ -22,11 +22,7 @@ export class TransactionToast {
   render() {
     return (
       <transaction-toast-wrapper wrapperId={`toast-${this.toastId}`} wrapperClass={this.wrapperClass}>
-        <transaction-toast-progress
-          key={this.toastId}
-          progressClass={this.transactionProgressState?.progressClass}
-          currentRemaining={this.transactionProgressState?.currentRemaining}
-        >
+        <transaction-toast-progress key={this.toastId} startTime={this.transactionProgressState?.startTime} endTime={this.transactionProgressState?.endTime}>
           <transaction-toast-content
             onDeleteToast={() => this.handleDeleteToast.emit(this.toastId)}
             processedTransactionsStatus={this.processedTransactionsStatus}